Refactor splitindex decision logic to own class
