+++ src/main/java/org/kairosdb/datastore/cassandra/CassandraDatastore.java
@@ -120,6 +120,8 @@ public class CassandraDatastore implements Datastore, KairosMetricReporter {
     private final StringKeyCache metricNameCache;
     private final StringKeyCache tagNameCache;
 
+    private final SplitIndexDecider splitIndexDecider;
+
     private final KairosDataPointFactory m_kairosDataPointFactory;
     private final LongDataPointFactory m_longDataPointFactory;
 
@@ -148,12 +150,15 @@ public class CassandraDatastore implements Datastore, KairosMetricReporter {
                               RowKeyCache rowKeyCache,
                               @Named(METRIC_NAME_CACHE) StringKeyCache metricNameCache,
                               @Named(TAG_NAME_CACHE) StringKeyCache tagNameCache,
+                              SplitIndexDecider splitIndexDecider,
                               Tracer tracer
     ) {
         m_cassandraConfiguration = cassandraConfiguration;
         this.rowKeyCache = rowKeyCache;
         this.metricNameCache = metricNameCache;
         this.tagNameCache = tagNameCache;
+        this.splitIndexDecider = splitIndexDecider;
+
 
         logger.warn("Setting tag index: {}", cassandraConfiguration.getIndexTagList());
         m_indexTagList = parseIndexTagList(cassandraConfiguration.getIndexTagList());
@@ -811,41 +816,22 @@ public class CassandraDatastore implements Datastore, KairosMetricReporter {
 
     // TODO remove when old getMatchingRowKeys is uncommented
     private List<DataPointsRowKey> getMatchingRowKeys(DatastoreMetricQuery query, int limit) {
-        // determine whether to use split index or not
-        String useSplitField = null;
-        Set<String> useSplitSet = new HashSet<>();
-
         Span span = tracer.buildSpan("query_index").start();
-
         try (Scope scope = tracer.scopeManager().activate(span, false)) {
             final List<String> indexTags = getIndexTags(query.getName());
             SetMultimap<String, String> filterTags = query.getTags();
-            if (useMetricAsSplitField(indexTags, filterTags)){
-                useSplitField = "metric";
-                useSplitSet = filterTags.get("key").stream().map((key) -> extractMetricsFromKey(key)).collect(Collectors.toSet());
-            } else {
-                for (String split : indexTags) {
-                    if (filterTags.containsKey(split)) {
-                        Set<String> currentSet = filterTags.get(split);
-                        final boolean currentSetIsSmaller = currentSet.size() < useSplitSet.size();
-                        final boolean currentSetIsNotEmpty = currentSet.size() > 0 && useSplitSet.isEmpty();
-                        final boolean currentSetHasNoWildcards = currentSet.stream().noneMatch(x -> x.contains("*") || x.contains("?"));
-                        if ((currentSetIsSmaller || currentSetIsNotEmpty) && currentSetHasNoWildcards) {
-                            useSplitSet = currentSet;
-                            useSplitField = split;
-                        }
-                    }
-                }
-            }
-            List<String> useSplit = new ArrayList<>(useSplitSet);
+            // Determine whether to use split index or not
+            SplitIndex splitIndex = splitIndexDecider.decideSplitIndex(indexTags, filterTags);
+            List<String> useSplit = new ArrayList<>(splitIndex.getSplitIndexValues());
+
             long startTime = calculateRowTimeRead(query.getStartTime());
             // Use write width here, as END time is upper bound for query and end with produces the bigger timestamp
             long endTime = calculateRowTimeWrite(query.getEndTime());
             logger.info("calculated: s={} e={}", startTime, endTime);
 
-            if (useSplitField != null && !"".equals(useSplitField) && useSplitSet.size() > 0) {
+            if (splitIndex.getSplitIndexField() != null && !"".equals(splitIndex.getSplitIndexField()) && splitIndex.getSplitIndexValues().size() > 0) {
                 span.setTag("type", "split");
-                return getMatchingRowKeysFromSplitIndex(query, useSplitField, useSplit, startTime, endTime, limit);
+                return getMatchingRowKeysFromSplitIndex(query, splitIndex.getSplitIndexField(), useSplit, startTime, endTime, limit);
             } else {
                 span.setTag("type", "global");
                 return getMatchingRowKeysFromRegularIndex(query, startTime, endTime, limit);
@@ -857,39 +843,6 @@ public class CassandraDatastore implements Datastore, KairosMetricReporter {
         } finally {
             span.finish();
         }
-    }
-
-    /* Rule when to use metric as split index key
-    - Only use metric if any value of key tag has wild cards
-    - If all other provided split index tags in the query do not have wildcards - Do not use metric as split index
-    - If metric does not have wildcards */
-    private boolean useMetricAsSplitField (List<String> indexTags, SetMultimap<String, String> filterTags) {
-        if (filterTags.containsKey("key")) {
-            Set<String> keyValues = filterTags.get("key");
-            boolean keyHasWildcards = keyValues.stream().anyMatch(x -> x.contains("*") || x.contains("?"));
-            if (keyHasWildcards && doAllSplitIndexTagsHaveWildCards(indexTags, filterTags)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private boolean doAllSplitIndexTagsHaveWildCards(List<String> indexTags, SetMultimap<String, String> filterTags) {
-        for (String splitTag: indexTags) {
-            if (filterTags.containsKey(splitTag)){
-                Set<String> tagValues = filterTags.get(splitTag);
-                if (tagValues.stream().noneMatch(x -> x.contains("*") || x.contains("?")) == true){
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-
-    private String extractMetricsFromKey(String key) {
-        String[] segments = key.split(".");
-        String metric = segments[segments.length - 1];
-        return metric.contains("*") || metric.contains("?") ? "" : metric;
     }
 
     private List<DataPointsRowKey> getMatchingRowKeysFromRegularIndex(DatastoreMetricQuery query,
