+++ src/test/java/org/kairosdb/datastore/cassandra/CassandraDatastoreTest.java
@@ -189,7 +189,7 @@ public class CassandraDatastoreTest extends DatastoreTestHelper {
         // TODO: test the caches being hit
         final StringKeyCache stringCache = mock(StringKeyCache.class);
         s_datastore = new CassandraDatastore(new CassandraClientImpl(cassandraConfig), cassandraConfig,
-                dataPointFactory, longDataPointFactory, mock(RowKeyCache.class), stringCache, stringCache, tracer);
+                dataPointFactory, longDataPointFactory, mock(RowKeyCache.class), stringCache, stringCache, new SplitIndexDecider(), tracer);
 
         System.out.println("Creating KairosDataStore");
         DatastoreTestHelper.s_datastore = new KairosDatastore(s_datastore,
